ConsoleUI.java: Add askValueWithPrompt method.
AskValueWithPrompt takes a String that it will prompt the user
with to get an integer from the user. Make CalculatorConsole use this
method and add final Strings to it for the prompts. Turn the askFirstValue
and askSecondValue private once again, as they're no longer needed to be
visible outside the class.